refactor: move credential card styles to a helper hook
